+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## NEXT
 * Stop muted accounts from sneaking into the feed via trending toots
-* Add `Toot.isDM()` instance method
+* Add `Toot.isDM()` instance method and filter option
 
 ## `0.9.1`
 * `DiversityFeedScorer` sorts by MD5 of id to get random but repeatable ordering so scores don't bounce around as incremental loads happen